My assistnant/requirements.txt
# Core dependencies
numpy>=1.20.0
sounddevice>=0.4.5
scipy>=1.7.0
webrtcvad>=2.0.10
pyttsx3>=2.90
requests>=2.26.0
python-dotenv>=0.19.0
llama-cpp-python>=0.2.0

# Speech recognition
SpeechRecognition>=3.8.1
pyaudio>=0.2.11
pvporcupine>=2.1.0

# Google API dependencies
google-auth-oauthlib>=0.4.6
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

# Notion API
notion-client>=1.0.0 